Website Version 1.9
Made absolutely major changes and changed basically the entire website. From layout to code, almost everything is new and it even takes on a newer design. It's only 1.9 because I am yet to update all the pages and so far only the home page is done. I will soon work on all pages and release v2.0 :D