Broke Everything :|
Broke everything because I couldn't stick to one thing at a time so was all over the place. I guess now is the best time to do one thing at a time and work on version 3.0. I'm tired and need to sleep so I'm just gonna have this commit and work on it at some other point and whenever I'm done with 3.0 I'll finally push the commits.